Initial Commit
This repository will contain all code towards the current
keystone-mfa implementation. Including the keystone plugin,
adjutant plugin and horizon plugin.

This is intened to work as MFA before a more long term solution
is introduced in Keystone later on. However most of this code
should be relevent with minor changes when that becomes avaliable.